fixup! GH-4952 - Introduce FedXConfig overrides for FedXRepositoryConfig
