+++ tools/federation/src/test/java/org/eclipse/rdf4j/federated/repository/FedXRepositoryConfigTest.java
@@ -12,6 +12,7 @@ package org.eclipse.rdf4j.federated.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.eclipse.rdf4j.model.util.Models.subject;
+import static org.junit.Assert.assertThat;
 
 import java.io.InputStream;
 import java.util.Optional;
@@ -27,6 +28,7 @@ import org.eclipse.rdf4j.repository.config.RepositoryConfigSchema;
 import org.eclipse.rdf4j.rio.RDFFormat;
 import org.eclipse.rdf4j.rio.Rio;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 public class FedXRepositoryConfigTest {
@@ -62,40 +64,6 @@ public class FedXRepositoryConfigTest {
 	}
 
 	@Test
-	public void testParseConfig_FedXConfig() throws Exception {
-		Model model = readConfig("/tests/rdf4jserver/config-withFedXConfig.ttl");
-
-		FedXRepositoryConfig config = new FedXRepositoryConfig();
-		config.parse(model, implNode(model));
-
-		FedXConfig fedXConf = config.getConfig();
-		Assertions.assertNotNull(fedXConf);
-
-		// Spot-check only: per-property testing is covered in FedXConfigParserTest
-		assertThat(fedXConf.getEnforceMaxQueryTime()).isEqualTo(42);
-		assertThat(fedXConf.isEnableMonitoring()).isTrue();
-		// A non-overridden option
-		assertThat(fedXConf.isDebugQueryPlan()).isFalse();
-
-		Assertions.assertNull(config.getMembers());
-		Assertions.assertNull(config.getDataConfig());
-	}
-
-	@Test
-	public void testParseConfig_FedXConfig_Overrides_Existing_Config() throws Exception {
-		Model model = readConfig("/tests/rdf4jserver/config-withFedXConfig.ttl");
-
-		FedXRepositoryConfig config = new FedXRepositoryConfig();
-		config.setConfig(new FedXConfig().withEnforceMaxQueryTime(33));
-		config.parse(model, implNode(model));
-
-		FedXConfig fedXConf = config.getConfig();
-		Assertions.assertNotNull(fedXConf);
-		// Read config should take precedence over pre-read state
-		assertThat(fedXConf.getEnforceMaxQueryTime()).isEqualTo(42);
-	}
-
-	@Test
 	public void testExport() throws Exception {
 		Model model = readConfig("/tests/rdf4jserver/config.ttl");
 
@@ -114,27 +82,6 @@ public class FedXRepositoryConfigTest {
 				.containsExactly("endpoint1", "endpoint2");
 	}
 
-	@Test
-	public void testExport_FedXConfig() throws Exception {
-		Model model = readConfig("/tests/rdf4jserver/config-withFedXConfig.ttl");
-
-		FedXRepositoryConfig config = new FedXRepositoryConfig();
-		config.parse(model, implNode(model));
-
-		// export into model
-		Model export = new TreeModel();
-		Resource implNode = config.export(export);
-
-		// Spot-check only: per-property testing is covered in FedXConfigParserTest
-		Optional<Resource> confNode = Models
-				.objectResource(export.getStatements(implNode, FedXRepositoryConfig.FEDX_CONFIG, null));
-		assertThat(confNode).hasValueSatisfying(node -> {
-			assertThat(Models
-					.objectLiteral(export.getStatements(node, FedXConfigParser.CONFIG_ENFORCE_MAX_QUERY_TIME, null)))
-					.hasValueSatisfying(v -> assertThat(v.intValue()).isEqualTo(42));
-		});
-	}
-
 	protected Model readConfig(String configResource) throws Exception {
 		try (InputStream in = FedXRepositoryConfigTest.class.getResourceAsStream(configResource)) {
 			return Rio.parse(in, "http://example.org/", RDFFormat.TURTLE);
@@ -145,4 +92,141 @@ public class FedXRepositoryConfigTest {
 		return subject(model.filter(null, RepositoryConfigSchema.REPOSITORYTYPE, null)).get();
 	}
 
+	@Nested
+	class FedXConfigParsing {
+
+		@Test
+		public void testParse() throws Exception {
+			Model model = readConfig("/tests/rdf4jserver/config-withFedXConfig.ttl");
+			FedXRepositoryConfig repoConfig = new FedXRepositoryConfig();
+
+			repoConfig.parse(model, implNode(model));
+			FedXConfig config = repoConfig.getConfig();
+
+			assertThat(config.getEnforceMaxQueryTime()).isEqualTo(1234);
+			assertThat(config.isEnableMonitoring()).isTrue();
+			assertThat(config.isLogQueryPlan()).isTrue();
+			assertThat(config.isDebugQueryPlan()).isTrue();
+			assertThat(config.isLogQueries()).isTrue();
+			assertThat(config.getSourceSelectionCacheSpec()).isEqualTo("spec-goes-here");
+		}
+
+		@Test
+		public void testParseConfigOverridesExistingConfig() throws Exception {
+			Model model = readConfig("/tests/rdf4jserver/config-withFedXConfig.ttl");
+			FedXRepositoryConfig repoConfig = new FedXRepositoryConfig();
+			repoConfig.setConfig(new FedXConfig().withEnforceMaxQueryTime(33));
+
+			repoConfig.parse(model, implNode(model));
+			FedXConfig config = repoConfig.getConfig();
+
+			assertThat(config.getEnforceMaxQueryTime()).isEqualTo(1234);
+			assertThat(config.isEnableMonitoring()).isTrue();
+			assertThat(config.isLogQueryPlan()).isTrue();
+			assertThat(config.isDebugQueryPlan()).isTrue();
+			assertThat(config.isLogQueries()).isTrue();
+			assertThat(config.getSourceSelectionCacheSpec()).isEqualTo("spec-goes-here");
+		}
+
+		@Test
+		public void testParseWithEmptyConfig() throws Exception {
+			Model model = readConfig("/tests/rdf4jserver/config.ttl");
+			FedXRepositoryConfig repoConfig = new FedXRepositoryConfig();
+			repoConfig.setConfig(new FedXConfig());
+
+			repoConfig.parse(model, implNode(model));
+			FedXConfig config = repoConfig.getConfig();
+
+			// expecting defaults
+			assertThat(config.getEnforceMaxQueryTime()).isEqualTo(30);
+			assertThat(config.isEnableMonitoring()).isFalse();
+			assertThat(config.isLogQueryPlan()).isFalse();
+			assertThat(config.isDebugQueryPlan()).isFalse();
+			assertThat(config.isLogQueries()).isFalse();
+			assertThat(config.getSourceSelectionCacheSpec()).isNull();
+		}
+
+		@Test
+		public void testExport() throws Exception {
+			Model model = readConfig("/tests/rdf4jserver/config-withFedXConfig.ttl");
+
+			FedXRepositoryConfig repoConfig = new FedXRepositoryConfig();
+			repoConfig.parse(model, implNode(model));
+
+			// export into model
+			Model export = new TreeModel();
+			Resource implNode = repoConfig.export(export);
+			Resource configNode = Models
+					.objectResource(export.getStatements(implNode, FedXRepositoryConfig.FEDX_CONFIG, null))
+					.orElse(null);
+			assertThat(configNode).isNotNull();
+
+			assertThat(export.filter(configNode, null, null)).hasSize(6);
+
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_ENFORCE_MAX_QUERY_TIME, null)))
+					.hasValueSatisfying(v -> assertThat(v.intValue()).isEqualTo(1234));
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_ENABLE_MONITORING, null)))
+					.hasValueSatisfying(v -> assertThat(v.booleanValue()).isTrue());
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_LOG_QUERY_PLAN, null)))
+					.hasValueSatisfying(v -> assertThat(v.booleanValue()).isTrue());
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_DEBUG_QUERY_PLAN, null)))
+					.hasValueSatisfying(v -> assertThat(v.booleanValue()).isTrue());
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_LOG_QUERIES, null)))
+					.hasValueSatisfying(v -> assertThat(v.booleanValue()).isTrue());
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_SOURCE_SELECTION_CACHE_SPEC,
+									null)))
+					.hasValueSatisfying(v -> assertThat(v.stringValue()).isEqualTo("spec-goes-here"));
+		}
+
+		@Test
+		public void testExportWithEmptyConfig() throws Exception {
+			FedXRepositoryConfig repoConfig = new FedXRepositoryConfig();
+			// Set to force export of defaults
+			repoConfig.setConfig(new FedXConfig());
+
+			// export into model
+			Model export = new TreeModel();
+			Resource implNode = repoConfig.export(export);
+			Resource configNode = Models
+					.objectResource(export.getStatements(implNode, FedXRepositoryConfig.FEDX_CONFIG, null))
+					.orElse(null);
+			assertThat(configNode).isNotNull();
+
+			// Note: 5 instead of 6 since CONFIG_SOURCE_SELECTION_CACHE_SPEC is null and thus should not be populated
+			assertThat(export.filter(configNode, null, null)).hasSize(5);
+
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_ENFORCE_MAX_QUERY_TIME, null)))
+					.hasValueSatisfying(v -> assertThat(v.intValue()).isEqualTo(30));
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_ENABLE_MONITORING, null)))
+					.hasValueSatisfying(v -> assertThat(v.booleanValue()).isFalse());
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_LOG_QUERY_PLAN, null)))
+					.hasValueSatisfying(v -> assertThat(v.booleanValue()).isFalse());
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_DEBUG_QUERY_PLAN, null)))
+					.hasValueSatisfying(v -> assertThat(v.booleanValue()).isFalse());
+			assertThat(
+					Models.objectLiteral(
+							export.getStatements(configNode, FedXRepositoryConfig.CONFIG_LOG_QUERIES, null)))
+					.hasValueSatisfying(v -> assertThat(v.booleanValue()).isFalse());
+		}
+	}
 }
