--- tools/federation/src/test/resources/tests/rdf4jserver/config-fedXConfig-only.ttl
@@ -1,16 +1,0 @@
-#
-# RDF4J FedXConfig overrides example
-#
-@prefix ex: <http://example.org/>.
-@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#>.
-@prefix rep: <http://www.openrdf.org/config/repository#>.
-@prefix fedx: <http://rdf4j.org/config/federation#>.
-
-# All overrides should have non-default values for test purposes
-ex:conf
-   fedx:enforceMaxQueryTime 1234 ;
-   fedx:enableMonitoring true ;
-   fedx:logQueryPlan true ;
-   fedx:debugQueryPlan true ;
-   fedx:logQueries true ;
-   fedx:sourceSelectionCacheSpec "spec-goes-here" .
